铁路（地下铁 地上铁 高架铁）生成器 更改开源许可证 修复打包时的奇奇怪怪错误

+++ makeRelease.js
@@ -2,21 +2,23 @@ const ar = require("archiver")
 const fs = require("fs")
 const cp = require('child_process')
 const path = require("path");
-function delPath(path) {//清除目录
-    var files = [];
-    if (fs.existsSync(path)) {
-        files = fs.readdirSync(path);
-        files.forEach(function (file, index) {
-            var curPath = path + "/" + file;
-            if (fs.statSync(curPath).isDirectory()) {
-                delPath(curPath);
-            } else {
-                fs.unlinkSync(curPath);
-            }
-        });
-        fs.rmdirSync(path);
+const requireMcfont = true
+function DelDirectory(dest) { //删除文件夹
+    if (fs.existsSync(dest) == false) {
+        return false;
     }
-};
+    var dirs = fs.readdirSync(dest);
+    dirs.forEach(function (item) {
+        var item_path = path.join(dest, item);
+        var temp = fs.statSync(item_path);
+        if (temp.isFile()) {
+            fs.unlink(item_path, (err) => {  });
+        } else if (temp.isDirectory()) {
+            DelDirectory(item_path);
+        }
+    });
+    fs.rmdir(dest, (err) => {  })
+}
 function CopyDirectory(src, dest) { //完整复制文件
     if (fs.existsSync(dest) == false) {
         fs.mkdir(dest, (e) => { });
@@ -35,21 +37,22 @@ function CopyDirectory(src, dest) { //完整复制文件
         }
     });
 }
-delPath(path.resolve("./release/"))
-fs.mkdir(path.resolve("./release/"), (e) => { })
-fs.mkdir(path.resolve("./release/item/"), (e) => { })
+DelDirectory(path.resolve("./release/"))
+fs.mkdir(path.resolve("./release/"), (err) => {  })
+fs.mkdir(path.resolve("./release/item/"), (err) => {})
 CopyDirectory(path.resolve("./packs/"), path.resolve("./release/item/"))
-delPath(path.resolve("./release/item/behaviors/scripts/"))
+DelDirectory(path.resolve("./release/item/behaviors/scripts/"))
 cp.execSync("webpack ./packs/behaviors/scripts/client/client.js -o ./release/item/behaviors/scripts/client")
-//非定制版的代码：fs.renameSync(path.resolve("./release/item/behaviors/scripts/client/main.js"), path.resolve("./release/item/behaviors/scripts/client/client.js"))
-//nc定制功能——防傻逼scriptingAPI的解码错误 {
-fs.writeFileSync(path.resolve("./release/item/behaviors/scripts/client/client.js"), fs.readFileSync(path.resolve("./release/item/behaviors/scripts/client/main.js"), { encoding: "utf-8" }).replace('"this is a debug message"', fs.readFileSync(path.resolve("./packs/behaviors/scripts/preset/font.real"), { encoding: "utf-8" })), { encoding: "utf-8" })
-fs.unlinkSync(path.resolve("./release/item/behaviors/scripts/client/main.js"))
-//}
+if (requireMcfont) {//nc定制功能——防傻逼scriptingAPI的解码错误 {
+    fs.writeFileSync(path.resolve("./release/item/behaviors/scripts/client/client.js"), fs.readFileSync(path.resolve("./release/item/behaviors/scripts/client/main.js"), { encoding: "utf-8" }).replace('"this is a debug message"', fs.readFileSync(path.resolve("./packs/behaviors/scripts/preset/font.real"), { encoding: "utf-8" })), { encoding: "utf-8" })
+    fs.unlinkSync(path.resolve("./release/item/behaviors/scripts/client/main.js"))
+} else {
+    fs.renameSync(path.resolve("./release/item/behaviors/scripts/client/main.js"), path.resolve("./release/item/behaviors/scripts/client/client.js"))
+}
 cp.execSync("webpack ./packs/behaviors/scripts/server/server.js -o ./release/item/behaviors/scripts/server")
 fs.renameSync(path.resolve("./release/item/behaviors/scripts/server/main.js"), path.resolve("./release/item/behaviors/scripts/server/server.js"))
 let releasezip = ar('zip', { zlib: { level: 9 } }) 
-fs.mkdir(path.resolve("./release/push/"), function (e) { })
+fs.mkdir(path.resolve("./release/push/"), (err) => { } )
 releasezip.pipe(fs.createWriteStream("./release/push/Release.zip"))
 releasezip.append(fs.createReadStream(path.resolve("./LICENSE")), { name: 'LICENSE' })
 releasezip.directory(path.resolve('./release/item/behaviors/'), 'behaviors')
